refact: rename package name in domain
