+++ post/src/main/kotlin/com/bibbi/domain/model/PostType.kt
@@ -1,4 +1,4 @@
-package com.bibbi.domain.entity
+package com.bibbi.domain.model
 
 import java.util.*
