+++ post/src/main/kotlin/com/bibbi/domain/model/RealEmoji.kt
@@ -1,4 +1,4 @@
-package com.bibbi.domain.entity
+package com.bibbi.domain.model
 
 data class RealEmoji(
     val id: String,
