+++ post/src/main/kotlin/com/bibbi/infrastructure/jpa/entity/CommentEntity.kt
@@ -1,7 +1,7 @@
 package com.bibbi.infrastructure.jpa.entity
 
 import com.bibbi.infrastructure.jpa.entity.BaseEntity
-import com.bibbi.domain.entity.Comment
+import com.bibbi.domain.model.Comment
 import jakarta.persistence.*
 import lombok.*
