+++ post/src/main/kotlin/com/bibbi/infrastructure/jpa/entity/PostEntity.kt
@@ -1,8 +1,8 @@
 package com.bibbi.infrastructure.jpa.entity
 
 import com.bibbi.infrastructure.jpa.entity.BaseEntity
-import com.bibbi.domain.entity.Post
-import com.bibbi.domain.entity.PostType
+import com.bibbi.domain.model.Post
+import com.bibbi.domain.model.PostType
 import jakarta.persistence.*
 import lombok.*
