+++ post/src/main/kotlin/com/bibbi/infrastructure/jpa/entity/ReactionEntity.kt
@@ -1,8 +1,8 @@
 package com.bibbi.infrastructure.jpa.entity
 
 import com.bibbi.infrastructure.jpa.entity.BaseEntity
-import com.bibbi.domain.entity.Emoji
-import com.bibbi.domain.entity.Reaction
+import com.bibbi.domain.model.Emoji
+import com.bibbi.domain.model.Reaction
 import jakarta.persistence.*
 
 @Entity(name = "reaction")
