+++ post/src/main/kotlin/com/bibbi/infrastructure/jpa/entity/RealEmojiEntity.kt
@@ -1,7 +1,7 @@
 package com.bibbi.infrastructure.jpa.entity
 
 import com.bibbi.infrastructure.jpa.entity.BaseEntity
-import com.bibbi.domain.entity.RealEmoji
+import com.bibbi.domain.model.RealEmoji
 import jakarta.persistence.*
 
 @Entity(name = "real_emoji")
